feat: complete "signup for early access" action even with login step
Earlier "signup for early access" button when pressed without logging in
would redirect to login and that would come back to the couse page and
the user need to click on the button again to complete the action. Fixed
it by adding a new endpoint to take care of that.